Nvidia_NeMo/requirements.txt
nvidia-nemo==1.20.0
pydub==0.25.1
ffmpeg-python==0.2.0
torch==2.1.2
torchaudio==2.1.2 